Make the windows VST callback use int32_t rather than long (to make it the same as the linux VST one); I think this is safe as the windows VST code is only built on 32-bit.
git-svn-id: svn://localhost/ardour2/branches/3.0@10754 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/session.h
@@ -739,10 +739,10 @@ class Session : public PBD::StatefulDestructible, public PBD::ScopedConnectionLi
 
 	/* VST support */
 
-	static long vst_callback (AEffect* effect,
-			long opcode,
-			long index,
-			long value,
+	static intptr_t vst_callback (AEffect* effect,
+			int32_t opcode,
+			int32_t index,
+			intptr_t value,
 			void* ptr,
 			float opt);
